Rename star rating and picker files
