+++ src/main/js/components/index.js
@@ -1,2 +1,2 @@
-export { default as StarRating } from "./rating";
-export { default as StarPicker } from "./starpicker";
+export { default as StarRating } from "./star-rating";
+export { default as StarPicker } from "./star-picker";
