+++ src/main/js/components/star-picker.js
