+++ src/main/js/components/star-rating.js
